Backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
gunicorn==20.1.0
psycopg2-binary==2.9.9prometheus-flask-exporter==0.23.0
